requirements.txt
Flask==2.3.3
gunicorn==21.2.0
torch>=2.0.0
transformers>=4.21.0
wordfreq>=3.0.0
Flask-Limiter>=3.5.0
python-dotenv>=1.0.0